+++ frontend/src/components/index.js
@@ -1,7 +1,13 @@
 import AppNavbar from './navbar'
 import AppRoutes from './routes'
+import CMSOMSCard from './cmsOMSCard'
+import ResponsivePlot from './responsivePlot'
+import Table from './table'
 
 export {
   AppNavbar,
-  AppRoutes
+  AppRoutes,
+  CMSOMSCard,
+  ResponsivePlot,
+  Table
 }
