+++ frontend/src/views/dataExplorer/histogram.js
@@ -11,8 +11,7 @@ import Button from 'react-bootstrap/Button'
 import { toast } from 'react-toastify'
 
 import API from '../../services/api'
-import ResponsivePlot from '../../components/responsivePlot'
-import CMSOMSCard from '../../components/cmsOMSCard'
+import { CMSOMSCard, ResponsivePlot } from '../../components'
 import { isNumericNonZero, isStringNonEmpty } from '../../utils/sanitizer'
 
 const Histogram = (props) => {
