+++ frontend/src/views/dataExplorer/runs.js
@@ -9,7 +9,7 @@ import Row from 'react-bootstrap/Row'
 import paginationFactory from 'react-bootstrap-table2-paginator'
 import { toast } from 'react-toastify'
 
-import Table from '../../components/table'
+import { Table } from '../../components'
 import API from '../../services/api'
 import { isNumericNonZero } from '../../utils/sanitizer'
